for publish

+++ app/Http/Controllers/ContactController.php
@@ -49,7 +49,8 @@ class ContactController extends Controller
         foreach($sections as $section) {
             $contacts = $section->contacts()->where('status', 'pending')->get();
             array_push($data, [
-                'section' => $section->name, 'contacts' => $contacts
+                'sectionName' => $section->name, 
+                'contacts' => $contacts
             ]);
         }
 
@@ -67,6 +68,23 @@ class ContactController extends Controller
         ], 200);
     }
 
+    public function myContacts() {
+        if (!auth('api')->user()){
+            return response()->json([
+                'message' => 'Unauthorized'
+            ], 401);
+        }
+
+        $user = User::find(auth('api')->user()->id);
+        
+        $contacts = $user->contacts;
+
+        return response()->json([
+            'data' => $contacts
+        ], 200);
+    }
+
+
 
     public function create(StoreContactRequest $request) {
         if (!auth('api')->user()) {
@@ -75,7 +93,14 @@ class ContactController extends Controller
             ], 401);
         }
 
-        $contact = Contact::create($request->all() + ['user_id' => auth('api')->user()->id, 'section_id' => auth('api')->user()->section_id]);
+        $contact = Contact::create([
+            'first_name' => $request->firstName,
+            'last_name'  => $request->lastName,
+            'mobile'     => $request->mobile,
+            'address'    => $request->address,
+            'user_id' => auth('api')->user()->id,
+            'section_id' => auth('api')->user()->section_id
+        ]);
 
         Log::create([
             'name' => auth('api')->user()->name,
@@ -102,6 +127,29 @@ class ContactController extends Controller
         return response()->json([
             'message' => 'Contact created successfully.',
             'id'      => $contact->id
+        ], 201);
+    }
+
+    public function edit($id) {
+
+        if (!auth('api')->user()) {
+            return response()->json([
+                'message' => 'Unauthorized'
+            ], 401);
+        }
+
+        $contact = Contact::find($id);
+
+        if (auth('api')->user()->role_id == 3 && auth('api')->user()->id != $contact->user_id) {
+            return response()->json([
+                'message' => 'Unauthorized'
+            ], 401);
+        }
+
+        // $contact->firstName = $contact->first_name;
+        // $contact->lastName = $contact->last_name;
+        return response()->json([
+            'data' => $contact
         ], 200);
     }
 
@@ -121,9 +169,15 @@ class ContactController extends Controller
             ], 401);
         }
 
+        if (request('counter') != $contact->counter) {
+            return response()->json([
+                'message' => 'This record is updated by another user, Please refresh the page.'
+            ], 422);
+        }
+
         if (in_array($id, ContactController::$process)) {
             return response()->json([
-                'message' => 'You can not update this contact because updating by another user.'
+                'message' => 'You can not update this contact because it is updating by another user.'
             ], 422);
         }
         
@@ -133,8 +187,8 @@ class ContactController extends Controller
 
         try {
             $contact = Contact::find($id);
-            $contact->first_name = request('first_name');
-            $contact->last_name = request('last_name');
+            $contact->first_name = request('firstName');
+            $contact->last_name = request('lastName');
             $contact->mobile = request('mobile');
             $contact->address = request('address');
             $contact->save();
@@ -142,11 +196,10 @@ class ContactController extends Controller
         } catch (\Excerption $e) {
             DB::rollback();
         }
-
-        $index = array_search($id, $process);
+        $index = array_search($id, ContactController::$process);
         if ($index) {
-            unset($process[$index]);
-            array_values($process);
+            unset(ContactController::$process[$index]);
+            array_values(ContactController::$process);
         }
 
         Log::create([
@@ -282,6 +335,9 @@ class ContactController extends Controller
         $ids = $request->ids;
         $joinedContact = $request->contact;
 
+        // making transaction.
+
+        
         if (auth('api')->user()->role_id != 1) {
             foreach($ids as $id) {
                 $contact = auth('api')->user()->contacts()
@@ -292,52 +348,41 @@ class ContactController extends Controller
                         'message' => 'Unauthorized'
                     ], 401);
                 }
-                if (in_array($id, ContactCotnroller::$process)) {
-                    return response()->json([
-                        'message' => 'You can not update this contact, because it is updating by another user.'
-                    ], 422);
-                }
             }
         }
 
         foreach($ids as $id) {
-            array_push(ContactController::$process, $id);
-        }
-
-        DB::beginTransaction();
-        
-        try {
-            foreach($ids as $id) {
-                $contact = Contact::find($id);
+            $contact = Contact::find($id);
+            if ($contact && $contact->order) {
                 $contact->order()->delete();
-                $contact->delete();
             }
-    
-            Contact::create($joinedContact + ['user_id' => auth('api')->user()->id, 'section_id' => auth('api')->user()->section_id]);
-            
-            unset(ContactController::$process);
-            ContactController::$process = [];
-
-            DB::commit();
-
-            Log::create([
-                'name' => auth('api')->user()->name,
-                'email' => auth('api')->user()->email,
-                'action' => 'joining muliple contacts.',
-                'table'  => 'contacts',
-                'role'   => auth('api')->user()->role->name
-            ]);
-    
-    
-            return response()->json([
-                'message' => 'Mergin contacts successfully.'
-            ], 200);
-        } catch(\Exception $e) {
-            DB::rollback();
-            return response()->json([
-                'message' => 'Internal error.'
-            ], 500);
+            // $contact->order()->delete();
+            if ($contact) {
+                $contact->delete();                
+            }
         }
+
+        Contact::create([
+            'first_name' => $joinedContact['first_name'],
+            'last_name'  => $joinedContact['last_name'],
+            'mobile'     => $joinedContact['mobile'],
+            'address'    => $joinedContact['address'],
+            'user_id' => auth('api')->user()->id, 
+            'section_id' => auth('api')->user()->section_id
+        ]);
+        
+
+        Log::create([
+            'name' => auth('api')->user()->name,
+            'email' => auth('api')->user()->email,
+            'action' => 'joining muliple contacts.',
+            'table'  => 'contacts',
+            'role'   => auth('api')->user()->role->name
+        ]);
+
+        return response()->json([
+            'message' => 'Merging contacts successfully.'
+        ], 200);
 
     }
